libsyntax: forbid visibility modifiers for enum variants
fixes #28433
