+++ src/test/compile-fail/issue-28433.rs
@@ -1,4 +1,4 @@
-// Copyright 2013 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2015 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,15 +8,12 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-struct A { pub i: isize }
-pub enum C { pub Variant }      //~ ERROR: unnecessary `pub`
-
-pub trait E {
-    fn foo(&self);
+enum bird {
+    pub duck, //~ ERROR: expected identifier, found keyword `pub`
+    goose
 }
 
-impl E for A {
-    pub fn foo(&self) {}             //~ ERROR: unnecessary visibility
-}
 
-fn main() {}
+fn main() {
+    let y = bird::goose;
+}
