+++ src/test/compile-fail/useless-pub.rs
@@ -1,4 +1,4 @@
-// Copyright 2012 The Rust Project Developers. See the COPYRIGHT
+// Copyright 2013 The Rust Project Developers. See the COPYRIGHT
 // file at the top-level directory of this distribution and at
 // http://rust-lang.org/COPYRIGHT.
 //
@@ -8,16 +8,14 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use zoo::bird::{duck, goose};
+struct A { pub i: isize }
 
-mod zoo {
-    pub enum bird {
-        pub duck, //~ ERROR: unnecessary `pub` visibility
-        goose
-    }
+pub trait E {
+    fn foo(&self);
 }
 
-
-fn main() {
-    let y = goose;
+impl E for A {
+    pub fn foo(&self) {}             //~ ERROR: unnecessary visibility
 }
+
+fn main() {}
